// src/models/Agora.ts
export interface AgoraDocument {
    agora_id: number;
    official_name: string;
    casual_name: string;
    link_to_document: string;
    authority: string;
    collections: string;
    most_recent_activity: "Enacted" | "Proposed" | "Defunct";
    most_recent_activity_date: string;
    proposed_date: string;
    annotated: boolean;
    validated: boolean;
    primarily_applies_to_the_government: boolean;
    primarily_applies_to_the_private_sector: boolean;
    short_summary: string;
    long_summary: string;
    tags: string | null;
    summaries_and_tags_may_include_unreviewed_machine_output: boolean;
    official_plaintext_retrieved: string | null;
    official_plaintext_source: string | null;
    official_plaintext_unavailable_infeasible: boolean;
    official_pdf_source: string | null;
    official_pdf_retrieved: string | null;
    number_of_segments_created: number;
    applications_agriculture_and_resource_extraction: boolean;
    applications_arts_sports_leisure_travel_and_lifestyle: boolean;
    applications_broadcasting_and_media_production: boolean;
    applications_business_services_and_analytics: boolean;
    applications_construction_and_field_services: boolean;
    applications_consumer_goods: boolean;
    applications_education: boolean;
    applications_energy_and_utilities: boolean;
    applications_finance_and_investment: boolean;
    applications_government_benefits_and_welfare: boolean;
    applications_government_judicial_and_law_enforcement: boolean;
    applications_government_military_and_public_safety: boolean;
    applications_government_other_applications_unspecified: boolean;
    applications_manufacturing_and_process_automation: boolean;
    applications_medicine_life_sciences_and_public_health: boolean;
    applications_networking_and_telecommunications: boolean;
    applications_sales_retail_and_customer_relations: boolean;
    applications_security: boolean;
    applications_transportation: boolean;
    harms_detrimental_content: boolean;
    harms_discrimination: boolean;
    harms_ecological_harm: boolean;
    harms_financial_loss: boolean;
    harms_harm_to_health_safety: boolean;
    harms_harm_to_infrastructure: boolean;
    harms_harm_to_property: boolean;
    harms_violation_of_civil_or_human_rights_including_privacy: boolean;
    incentives_access_to_business_opportunities: boolean;
    incentives_civil_liability: boolean;
    incentives_criminal_liability: boolean;
    incentives_fines: boolean;
    incentives_imprisonment: boolean;
    incentives_subsidies: boolean;
    risk_factors_bias: boolean;
    risk_factors_interpretability_and_explainability: boolean;
    risk_factors_privacy: boolean;
    risk_factors_reliability: boolean;
    risk_factors_reliability_robustness: boolean;
    risk_factors_safety: boolean;
    risk_factors_security: boolean;
    risk_factors_security_cybersecurity: boolean;
    risk_factors_security_dissemination: boolean;
    risk_factors_transparency: boolean;
    strategies_convening: boolean;
    strategies_disclosure: boolean;
    strategies_disclosure_about_evaluation: boolean;
    strategies_disclosure_about_incidents: boolean;
    strategies_disclosure_about_inputs: boolean;
    strategies_disclosure_accuracy_thereof: boolean;
    strategies_disclosure_in_deployment: boolean;
    strategies_disclosure_in_standard_form: boolean;
    strategies_evaluation: boolean;
    strategies_evaluation_adversarial_testing: boolean;
    strategies_evaluation_conformity_assessment: boolean;
    strategies_evaluation_external_auditing: boolean;
    strategies_evaluation_impact_assessment: boolean;
    strategies_evaluation_post_market_monitoring: boolean;
    strategies_governance_development: boolean;
    strategies_government_study_or_report: boolean;
    strategies_government_support: boolean;
    strategies_government_support_ai_workforce_related: boolean;
    strategies_government_support_for_r_d: boolean;
    strategies_input_controls: boolean;
    strategies_input_controls_compute_circulation: boolean;
    strategies_input_controls_compute_use: boolean;
    strategies_input_controls_data_circulation: boolean;
    strategies_input_controls_data_use: boolean;
    strategies_licensing_registration_and_certification: boolean;
    strategies_new_institution: boolean;
    strategies_performance_requirements: boolean;
    strategies_pilots_and_testbeds: boolean;
    strategies_tiering: boolean;
    strategies_tiering_tiering_based_on_domain_of_application: boolean;
    strategies_tiering_tiering_based_on_generality: boolean;
    strategies_tiering_tiering_based_on_impact: boolean;
    strategies_tiering_tiering_based_on_inputs: boolean;
    strategies_tiering_tiering_based_on_planning_ability: boolean;
    jurisdiction: string | null;
    parent_authority: string | null;
  }